+++ packages/wallets/ambire/src/index.ts
@@ -2,11 +2,16 @@ import { WalletAdapterType } from '@reef-knot/types';
 import { getWalletConnectConnector } from '@reef-knot/wallets-helpers';
 import WalletIcon from './icons/ambire.svg';
 
-export const Ambire: WalletAdapterType = ({ rpc, walletconnectProjectId }) => ({
+export const Ambire: WalletAdapterType = ({
+  rpc,
+  walletconnectProjectId,
+  chains,
+}) => ({
   walletName: 'Ambire',
   walletId: 'ambire',
   icon: WalletIcon,
   connector: getWalletConnectConnector({
+    chains,
     rpc,
     noMobileLinks: true,
     projectId: walletconnectProjectId,
@@ -14,6 +19,7 @@ export const Ambire: WalletAdapterType = ({ rpc, walletconnectProjectId }) => ({
   walletconnectExtras: {
     connectionViaURI: {
       connector: getWalletConnectConnector({
+        chains,
         rpc,
         qrcode: false,
         projectId: walletconnectProjectId,