+++ packages/wallets/blockchaincom/src/index.ts
@@ -6,6 +6,7 @@ import WalletIconInverted from './icons/blockchaincom-inverted.svg';
 export const Blockchaincom: WalletAdapterType = ({
   rpc,
   walletconnectProjectId,
+  chains,
 }) => ({
   walletName: 'Blockchain.com',
   walletId: 'blockchaincom',
@@ -14,6 +15,7 @@ export const Blockchaincom: WalletAdapterType = ({
     dark: WalletIconInverted,
   },
   connector: getWalletConnectConnector({
+    chains,
     rpc,
     projectId: walletconnectProjectId,
     noMobileLinks: true,