+++ packages/wallets/walletconnect/src/index.ts
@@ -5,11 +5,13 @@ import WalletIcon from './icons/wallet-connect-circle.svg';
 export const WalletConnect: WalletAdapterType = ({
   rpc,
   walletconnectProjectId,
+  chains,
 }) => ({
   walletName: 'WalletConnect',
   walletId: 'walletconnect',
   icon: WalletIcon,
   connector: getWalletConnectConnector({
+    chains,
     rpc,
     projectId: walletconnectProjectId,
   }),