+++ packages/connect-wallet-modal/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @reef-knot/connect-wallet-modal
 
+## 1.4.0
+
+### Minor Changes
+
+- Explicitly pass chains to wagmi connectors, update wagmi to v0.12.16
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 1.3.1
 
 ### Patch Changes